requirements.txt
# labelme
# labelme2yolo
pyttsx3
ultralytics
opencv-python
Flask
pillow
flask_cors
# torch